add test + clippy
